[2.1.1] ActionBar fixed: support 1.13 & Minor bug fixed
* Update ActionBarAPI to support 1.13
* Fix '/setspawn' prefix bug that not same with other portal (didn't use Prefix.portal)
* Update '/sethome' message
+ '/pondjaprivateserver' to toggle private server feature (Note: Only avalible for me only ;w;)
+ '/unloadchunk' to unload unnecessary chunk (this command is for debug right now, not recommand to use)
* Remove unnecessary code in '/free'
+ Add Ping Display on PlayerListTab
* Change WarpPlate message from "Hold" to "Press"
* Change Shop system's price calculation from double to long
* Fix FreeItem GUI emerald block and redstone block didn't display its message.
- Remove StringTest on OnPlayerMovement (This is a human-error, I don't even know when I wrote it here. LOL)
- Disabled DropItemBeforeChat system for temporary.

+++ src/me/palapon2545/SMDMain/Function/Shop.java
@@ -47,6 +47,7 @@ public class Shop {
 		File userdata = new File(StockInt.pluginDir, File.separator + "PlayerDatabase/" + playerName);
 		File f = new File(userdata, File.separator + "config.yml");
 		FileConfiguration playerData = YamlConfiguration.loadConfiguration(f);
+		
 		Material l = Material.getMaterial(item.toUpperCase());
 
 		if (l != null) {
@@ -62,7 +63,7 @@ public class Shop {
 
 			// can sell
 			if (sellCount > 0) {
-				double gotPrice = ((double) price / (double) amount) * sellCount;
+				long gotPrice = (long) (((double) price / (double) amount) * sellCount);
 				try {
 					playerData.set("money", money + gotPrice);
 					playerData.save(f);
